add bigquery import script
