+++ src/router.js
@@ -5,7 +5,6 @@ import data from './routes/bot_data.js';
 import db from './routes/db.js';
 import support from './routes/support.js';
 import webhook from './routes/webhook.js';
-import bigquery from './routes/big_query.js';
 
 const router = Router();
 
@@ -15,6 +14,5 @@ router.use('/data', data);
 router.use('/db', db);
 router.use('/support', support);
 router.use('/webhook', webhook);
-router.use('/big-query', bigquery);
 
 export default router;
